dividing cart in components
